add sync_item_dummy compilation unit
