+++ test/unittests/CMakeLists.txt
@@ -198,6 +198,7 @@ if (BUILD_SERVER)
        ${CVMFS_SOURCE_DIR}/swissknife_assistant.cc
        ${CVMFS_SOURCE_DIR}/swissknife_lease_curl.cc
        ${CVMFS_SOURCE_DIR}/sync_item.cc
+       ${CVMFS_SOURCE_DIR}/sync_item_dummy.cc
        ${CVMFS_SOURCE_DIR}/sync_item_tar.cc
        ${CVMFS_SOURCE_DIR}/sync_mediator.cc
        ${CVMFS_SOURCE_DIR}/sync_union.cc
@@ -502,6 +503,7 @@ set(CVMFS_UNITTEST_SOURCES
   ${CVMFS_SOURCE_DIR}/swissknife_lease_json.cc
   ${CVMFS_SOURCE_DIR}/swissknife_lease_curl.cc
   ${CVMFS_SOURCE_DIR}/sync_item.cc
+  ${CVMFS_SOURCE_DIR}/sync_item_dummy.cc
   ${CVMFS_SOURCE_DIR}/sync_item_tar.cc
   ${CVMFS_SOURCE_DIR}/sync_mediator.cc
   ${CVMFS_SOURCE_DIR}/sync_union.cc
